refactor: Extract StreamingContext into streaming_context.py
Split the 1,304-line streaming.py monolith: move StreamingContext
dataclass (207 lines) into dedicated streaming_context.py. The original
module re-exports StreamingContext for backwards compatibility.
streaming.py drops to 1,121 lines focused on event construction and
iterators.
